work api

+++ src/main/java/bfg/backend/dto/request/modulePlace/ModulePlace.java
@@ -1,6 +1,6 @@
 package bfg.backend.dto.request.modulePlace;
 
-public record ModulePlace(Integer idUser,
+public record ModulePlace(Long idUser,
                           Integer typeModule,
                           Integer x,
                           Integer y,
